change version fetching
importing from pkg_resources is painfully slow and I am not willing to pay that price for a version that is hardly ever interesting.

The inbuilt argparse version thing is nice but needs the version already fetched, so switching to a simpler not quite so nice way that gets rid if the extra ~300ms import time on startup that make it feel so sluggish.